+++ Changelog.md
@@ -1,5 +1,6 @@
 # Changes
 
+* 2016-11-30 displaythree command
 * 2016-11-30 seevar, seefuns, varref and funref, ie, thenextfigure, typeref, metavar commands
 * 2016-11-30 issue and endissue command - noop for now, let's implement them whenever we actually render list of issues
 * 2016-11-30 \term, \newterm commands, begin/end subsubsubsubsubsubsubsection commands in bulk